Etrade_fetch/requirements.txt
pandas~=2.3.3
sqlalchemy~=2.0.44
requests~=2.32.5
tqdm~=4.67.1
python-dotenv~=1.0.1
